AbstractClassPathRepository
