+++ src/main/java/org/apache/bcel/util/MemorySensitiveClassPathRepository.java
@@ -17,13 +17,10 @@
  */
 package org.apache.bcel.util;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.lang.ref.SoftReference;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.apache.bcel.classfile.ClassParser;
 import org.apache.bcel.classfile.JavaClass;
 
 /**
@@ -33,13 +30,12 @@ import org.apache.bcel.classfile.JavaClass;
  *
  * @see org.apache.bcel.Repository
  */
-public class MemorySensitiveClassPathRepository implements Repository {
+public class MemorySensitiveClassPathRepository extends AbstractClassPathRepository {
 
-    private ClassPath _path = null;
     private final Map<String, SoftReference<JavaClass>> _loadedClasses = new HashMap<>(); // CLASSNAME X JAVACLASS
 
     public MemorySensitiveClassPathRepository(final ClassPath path) {
-        this._path = path;
+        super(path);
     }
 
     /**
@@ -47,6 +43,7 @@ public class MemorySensitiveClassPathRepository implements Repository {
      */
     @Override
     public void storeClass(final JavaClass clazz) {
+        // Not calling super.storeClass because this subclass maintains the mapping.
         _loadedClasses.put(clazz.getClassName(), new SoftReference<>(clazz));
         clazz.setRepository(this);
     }
@@ -69,96 +66,6 @@ public class MemorySensitiveClassPathRepository implements Repository {
             return null;
         }
         return ref.get();
-    }
-
-    /**
-     * Find a JavaClass object by name. If it is already in this Repository, the Repository version is returned. Otherwise, the Repository's classpath is
-     * searched for the class (and it is added to the Repository if found).
-     *
-     * @param className
-     *            the name of the class
-     * @return the JavaClass object
-     * @throws ClassNotFoundException
-     *             if the class is not in the Repository, and could not be found on the classpath
-     */
-    @Override
-    public JavaClass loadClass(String className) throws ClassNotFoundException {
-        if ((className == null) || className.isEmpty()) {
-            throw new IllegalArgumentException("Invalid class name " + className);
-        }
-        className = className.replace('/', '.'); // Just in case, canonical form
-        final JavaClass clazz = findClass(className);
-        if (clazz != null) {
-            return clazz;
-        }
-        try {
-            return loadClass(_path.getInputStream(className), className);
-        } catch (final IOException e) {
-            throw new ClassNotFoundException("Exception while looking for class " + className + ": " + e, e);
-        }
-    }
-
-    /**
-     * Find the JavaClass object for a runtime Class object. If a class with the same name is already in this Repository, the Repository version is returned.
-     * Otherwise, getResourceAsStream() is called on the Class object to find the class's representation. If the representation is found, it is added to the
-     * Repository.
-     *
-     * @see Class
-     * @param clazz
-     *            the runtime Class object
-     * @return JavaClass object for given runtime class
-     * @throws ClassNotFoundException
-     *             if the class is not in the Repository, and its representation could not be found
-     */
-    @Override
-    public JavaClass loadClass(final Class<?> clazz) throws ClassNotFoundException {
-        final String className = clazz.getName();
-        final JavaClass repositoryClass = findClass(className);
-        if (repositoryClass != null) {
-            return repositoryClass;
-        }
-        String name = className;
-        final int i = name.lastIndexOf('.');
-        if (i > 0) {
-            name = name.substring(i + 1);
-        }
-        JavaClass cls = null;
-        try (InputStream clsStream = clazz.getResourceAsStream(name + ".class")) {
-            return cls = loadClass(clsStream, className);
-        } catch (final IOException e) {
-            return cls;
-        }
-
-    }
-
-    private JavaClass loadClass(final InputStream is, final String className) throws ClassNotFoundException {
-        try {
-            if (is != null) {
-                final ClassParser parser = new ClassParser(is, className);
-                final JavaClass clazz = parser.parse();
-                storeClass(clazz);
-                return clazz;
-            }
-        } catch (final IOException e) {
-            throw new ClassNotFoundException("Exception while looking for class " + className + ": " + e, e);
-        } finally {
-            if (is != null) {
-                try {
-                    is.close();
-                } catch (final IOException e) {
-                    // ignored
-                }
-            }
-        }
-        throw new ClassNotFoundException("SyntheticRepository could not load " + className);
-    }
-
-    /**
-     * ClassPath associated with the Repository.
-     */
-    @Override
-    public ClassPath getClassPath() {
-        return _path;
     }
 
     /**
